+++ src/main/java/com/visiontracker/challengeTrackerApplication/repositories/ProgramRepository.java
@@ -10,10 +10,10 @@ import org.springframework.web.bind.annotation.CrossOrigin;
 import java.util.List;
 
 @CrossOrigin(origins = "http://localhost:4200")
-public interface ProgramRepository extends CrudRepository<Program, Integer> {
+public interface ProgramRepository extends CrudRepository<Program, Long> {
 
     @Query("select p from Program p where p.programId = :programId")
-    public Program findProgramById(@Param("programId") Integer programId);
+    public Program findProgramById(@Param("programId") Long programId);
 
     @Query("select p from Program p")
     public List<Program> findAll();
