+++ src/main/java/com/visiontracker/challengeTrackerApplication/repositories/UserRepository.java
@@ -6,16 +6,18 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.web.bind.annotation.CrossOrigin;
 
+import java.sql.SQLException;
+import java.sql.SQLIntegrityConstraintViolationException;
 import java.util.List;
 
 @CrossOrigin(origins = "http://localhost:4200")
-public interface UserRepository extends CrudRepository<User, Integer> {
+public interface UserRepository extends CrudRepository<User, Long> {
 
     @Query("SELECT u FROM User u WHERE u.username = :username")
     public User findUserByUsername(@Param("username") String username);
 
     @Query("select u from User u where u.userId = :userId")
-    public User findUserById(@Param("userId") Integer userId);
+    public User findUserById(@Param("userId") Long userId);
 
     @Query("select u from User u")
     public List<User> findAll();
